# Server/api/openapi.yaml
...
openapi: "3.1.0"
info:
  title: "Inlay API"
  version: "2.0.0"
  description: "Shared transport contract for the Swift client and TypeScript server. Optional properties are omitted rather than encoded as null. Dates use whole-second ISO8601 UTC timestamps."
servers:
  - url: "http://localhost:8391"
security:
  - bearerAuth: []
...
paths:
...
  "/v1/preferences":
    get:
      operationId: "getPreferences"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/PreferencesSnapshot"
        default:
          "$ref": "#/components/responses/APIError"
    put:
      operationId: "updatePreferences"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/PreferencesSnapshot"
        default:
          "$ref": "#/components/responses/APIError"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/PreferencesSnapshot"
        x-maximumBytes: 262144
  "/v1/generations":
    get:
      operationId: "listGenerations"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/GenerationPage"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        - name: "limit"
          in: "query"
          required: false
          schema:
            type: "integer"
            minimum: 1
            maximum: 100
            default: 50
        - name: "before"
          in: "query"
          required: false
          schema:
            type: "string"
        - name: "source"
          in: "query"
          required: false
          schema:
            type: "string"
            enum:
              - "inlay"
              - "wispr-flow"
    post:
      operationId: "createGeneration"
      responses:
        "201":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/GenerationRecord"
        default:
          "$ref": "#/components/responses/APIError"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/CreateGenerationRequest"
        x-maximumBytes: 262144
  "/v1/generations/{id}":
    get:
      operationId: "getGeneration"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/GenerationRecord"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
    delete:
      operationId: "deleteGeneration"
      responses:
        "204":
          description: "Success"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
  "/v1/generations/{id}/audio/{kind}":
    post:
      operationId: "appendAudio"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/AudioChunkReceipt"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
        - name: "kind"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/AudioKind"
        - name: "sequence"
          in: "query"
          required: true
          schema:
            type: "integer"
            minimum: 0
            maximum: 4095
        - name: "sampleRate"
          in: "query"
          required: true
          schema:
            type: "integer"
            minimum: 8000
            maximum: 192000
        - name: "channels"
          in: "query"
          required: true
          schema:
            type: "integer"
            minimum: 1
            maximum: 8
      requestBody:
        required: true
        content:
          "application/octet-stream":
            schema:
              type: "string"
              format: "binary"
        description: "Raw interleaved little-endian float32 PCM. Chunks must contain whole frames. Repeated sequences must have identical bytes."
        x-maximumBytes: 1048576
  "/v1/generations/{id}/finish":
    post:
      operationId: "finishGeneration"
      responses:
        "202":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/GenerationRecord"
        default:
          "$ref": "#/components/responses/APIError"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/FinishGenerationRequest"
        x-maximumBytes: 262144
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
  "/v1/generations/{id}/cancel":
    post:
      operationId: "cancelGeneration"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/GenerationRecord"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
...
  "/v1/generations/{id}/events":
    get:
      operationId: "generationEvents"
      responses:
        "200":
          description: "Success"
          content:
            "application/x-ndjson":
              schema:
                type: "string"
              x-recordSchema:
                "$ref": "#/components/schemas/GenerationRecord"
        default:
          "$ref": "#/components/responses/APIError"
      description: "A bounded NDJSON stream of complete GenerationRecord values, with the latest record repeated as a two-second heartbeat. Ends after a terminal record."
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
  "/v1/generations/{id}/artifacts/{filename}":
    get:
      operationId: "downloadArtifact"
      responses:
        "200":
          description: "Success"
          content:
            "application/octet-stream":
              schema:
                type: "string"
                format: "binary"
            "audio/wav":
              schema:
                type: "string"
                format: "binary"
            "image/png":
              schema:
                type: "string"
                format: "binary"
            "application/json":
              schema:
                type: "string"
                format: "binary"
            "text/plain":
              schema:
                type: "string"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
        - name: "filename"
          in: "path"
          required: true
          schema:
            type: "string"
            enum:
              - "inference.wav"
              - "original.wav"
              - "metadata.json"
              - "transcript.txt"
              - "source.json"
              - "source.wav"
              - "opus.json"
              - "screenshot.png"
              - "built-in-audio.bin"
  "/v1/imports/wispr-flow/known":
    post:
      operationId: "knownWisprFlowIDs"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/WisprFlowKnownIDsResponse"
        default:
          "$ref": "#/components/responses/APIError"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/WisprFlowKnownIDsRequest"
        x-maximumBytes: 262144
  "/v1/imports/wispr-flow":
    post:
      operationId: "beginWisprFlowImport"
      responses:
        "201":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/WisprFlowImportSession"
        default:
          "$ref": "#/components/responses/APIError"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/WisprFlowImportRequest"
        x-maximumBytes: 262144
  "/v1/imports/wispr-flow/dictionary":
    put:
      operationId: "archiveWisprFlowDictionary"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/WisprFlowDictionaryArchiveReceipt"
        default:
          "$ref": "#/components/responses/APIError"
      requestBody:
        required: true
        content:
          "application/json":
            schema: {}
        x-maximumBytes: 8388608
        description: "Original Wispr Flow dictionary JSON archive. This does not change the active personal dictionary."
  "/v1/imports/wispr-flow/{id}/artifacts/{filename}":
    put:
      operationId: "uploadWisprFlowArtifact"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/WisprFlowArtifactReceipt"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
        - name: "filename"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/WisprFlowArtifactName"
      requestBody:
        required: true
        content:
          "application/octet-stream":
            schema:
              type: "string"
              format: "binary"
          "application/json":
            schema:
              type: "string"
              format: "binary"
          "audio/wav":
            schema:
              type: "string"
              format: "binary"
          "image/png":
            schema:
              type: "string"
              format: "binary"
        x-maximumBytes: 8388608
  "/v1/imports/wispr-flow/{id}/complete":
    post:
      operationId: "completeWisprFlowImport"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/WisprFlowImportResult"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
  "/v1/imports/wispr-flow/{id}":
    delete:
      operationId: "cancelWisprFlowImport"
      responses:
        "204":
          description: "Success"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
components:
  schemas:
    UUID:
      type: "string"
      format: "uuid"
    GenerationMode:
      type: "string"
      enum:
        - "dictation"
        - "test"
        - "file"
    GenerationStatus:
      type: "string"
      enum:
        - "receiving"
        - "queued"
        - "transcribing"
        - "proofreading"
        - "completed"
        - "failed"
        - "cancelled"
    AudioKind:
      type: "string"
      enum:
        - "inference"
        - "original"
    WisprFlowArtifactName:
      type: "string"
      enum:
        - "source.json"
        - "source.wav"
        - "opus.json"
        - "screenshot.png"
        - "built-in-audio.bin"
    WisprFlowImportOutcome:
      type: "string"
      enum:
        - "imported"
        - "enriched"
        - "skipped"
        - "partial"
...
    DeviceIdentity:
      type: "object"
      additionalProperties: false
      properties:
        id:
          type: "string"
        name:
          type: "string"
      required:
        - "id"
        - "name"
    ServerPreferences:
      type: "object"
      additionalProperties: false
      properties:
        language:
          description: "Proofreading language hint. Parakeet recognizes speech automatically; legacy codes remain accepted for stored history."
          type: "string"
          enum:
            - "en"
            - "auto"
            - "es"
            - "fr"
            - "de"
            - "it"
            - "pt"
            - "nl"
            - "ja"
            - "zh"
            - "ko"
            - "hi"
            - "ar"
            - "pl"
            - "ru"
            - "uk"
            - "sv"
        proofreadingPrompt:
          type: "string"
          minLength: 1
          description: "Missing values use the built-in cleanup prompt; maximum UTF-8 size is 4096 bytes."
        vocabulary:
          type: "string"
          description: "Maximum UTF-8 size is 16384 bytes."
        dictionary:
          "$ref": "#/components/schemas/PersonalDictionary"
        textCorrectionEnabled:
          type: "boolean"
        keepOriginalAudio:
          type: "boolean"
      required:
        - "language"
        - "vocabulary"
        - "dictionary"
        - "textCorrectionEnabled"
        - "keepOriginalAudio"
    PreferencesSnapshot:
      type: "object"
      additionalProperties: false
      properties:
        revision:
          type: "integer"
          minimum: 0
          maximum: 9007199254740991
        preferences:
          "$ref": "#/components/schemas/ServerPreferences"
      required:
        - "revision"
        - "preferences"
...
    CreateGenerationRequest:
      type: "object"
      additionalProperties: false
      properties:
        requestID:
          "$ref": "#/components/schemas/UUID"
        device:
          "$ref": "#/components/schemas/DeviceIdentity"
        mode:
          "$ref": "#/components/schemas/GenerationMode"
      required:
        - "requestID"
        - "device"
        - "mode"
...
    AudioChunkReceipt:
      type: "object"
      additionalProperties: false
      properties:
        nextSequence:
          type: "integer"
          minimum: 0
          maximum: 9007199254740991
        frameCount:
          type: "integer"
          minimum: 0
          maximum: 9007199254740991
          format: "int64"
      required:
        - "nextSequence"
        - "frameCount"
    FinishGenerationRequest:
      type: "object"
      additionalProperties: false
      properties:
        inferenceFrames:
          type: "integer"
          minimum: 0
          maximum: 9007199254740991
          format: "int64"
        originalFrames:
          type: "integer"
          minimum: 0
          maximum: 9007199254740991
          format: "int64"
        continuationID:
          "$ref": "#/components/schemas/UUID"
      required:
        - "inferenceFrames"
    AudioArtifact:
      type: "object"
      additionalProperties: false
      properties:
        filename:
          type: "string"
        sampleRate:
          type: "integer"
          minimum: 1
          maximum: 9007199254740991
        channels:
          type: "integer"
          minimum: 1
          maximum: 9007199254740991
        frameCount:
          type: "integer"
          minimum: 0
          maximum: 9007199254740991
          format: "int64"
        byteCount:
          type: "integer"
          minimum: 0
          maximum: 9007199254740991
          format: "int64"
        encoding:
          type: "string"
      required:
        - "filename"
        - "sampleRate"
        - "channels"
        - "frameCount"
        - "byteCount"
        - "encoding"
    ModelProvenance:
      type: "object"
      additionalProperties: false
      properties:
        modelID:
          type: "string"
        modelSHA256:
          type: "string"
        backend:
          type: "string"
        engineVersion:
          type: "string"
        processingSeconds:
          type: "number"
          minimum: 0
      required:
        - "modelID"
        - "backend"
...
    ModelHintUsage:
      type: "object"
      additionalProperties: false
      properties:
        includedTerms:
          type: "array"
          items:
            type: "string"
        omittedTerms:
          type: "array"
          items:
            type: "string"
        tokenCount:
          type: "integer"
          minimum: 0
          maximum: 9007199254740991
        tokenBudget:
          type: "integer"
          minimum: 0
          maximum: 9007199254740991
      required:
        - "includedTerms"
        - "omittedTerms"
...
    WisprFlowImportRequest:
      type: "object"
      additionalProperties: false
      properties:
        sourceID:
          "$ref": "#/components/schemas/UUID"
        createdAt:
          type: "string"
          format: "date-time"
        sourceStatus:
          type: "string"
        finalText:
          type: "string"
        rawText:
          type: "string"
        durationSeconds:
          type: "number"
          minimum: 0
        variantNames:
          type: "array"
          items:
            type: "string"
          maxItems: 32
        artifacts:
          type: "array"
          items:
            "$ref": "#/components/schemas/WisprFlowArtifactManifest"
          minItems: 1
          maxItems: 4
        unarchivedArtifacts:
          type: "array"
          items:
            "$ref": "#/components/schemas/WisprFlowArtifactManifest"
      required:
        - "sourceID"
        - "createdAt"
        - "finalText"
        - "rawText"
        - "variantNames"
        - "artifacts"
    WisprFlowImportSession:
      type: "object"
      additionalProperties: false
      properties:
        id:
          "$ref": "#/components/schemas/UUID"
      required:
        - "id"
    WisprFlowArtifactReceipt:
      type: "object"
      additionalProperties: false
      properties:
        filename:
          "$ref": "#/components/schemas/WisprFlowArtifactName"
        byteCount:
          type: "integer"
          minimum: 0
          maximum: 8388608
      required:
        - "filename"
        - "byteCount"
    WisprFlowKnownIDsRequest:
      type: "object"
      additionalProperties: false
      properties:
        sourceIDs:
          type: "array"
          items:
            "$ref": "#/components/schemas/UUID"
          maxItems: 10000
      required:
        - "sourceIDs"
    WisprFlowKnownIDsResponse:
      type: "object"
      additionalProperties: false
      properties:
        knownSourceIDs:
          type: "array"
          items:
            "$ref": "#/components/schemas/UUID"
      required:
        - "knownSourceIDs"
    WisprFlowDictionaryArchiveReceipt:
      type: "object"
      additionalProperties: false
      properties:
        byteCount:
          type: "integer"
          minimum: 0
          maximum: 8388608
        sha256:
          type: "string"
      required:
        - "byteCount"
        - "sha256"
...
    GenerationRecord:
      type: "object"
      additionalProperties: false
      properties:
        schemaVersion:
          type: "integer"
          minimum: 0
          maximum: 9007199254740991
        id:
          "$ref": "#/components/schemas/UUID"
        requestID:
          "$ref": "#/components/schemas/UUID"
        device:
          "$ref": "#/components/schemas/DeviceIdentity"
        mode:
          "$ref": "#/components/schemas/GenerationMode"
        status:
          "$ref": "#/components/schemas/GenerationStatus"
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"
        settings:
          "$ref": "#/components/schemas/PreferencesSnapshot"
        inferenceAudio:
          "$ref": "#/components/schemas/AudioArtifact"
        originalAudio:
          "$ref": "#/components/schemas/AudioArtifact"
        rawText:
          type: "string"
        finalText:
          type: "string"
        insertionText:
          type: "string"
        previewText:
          type: "string"
        detectedLanguage:
          type: "string"
        speech:
          "$ref": "#/components/schemas/ModelProvenance"
        proofreading:
          "$ref": "#/components/schemas/ModelProvenance"
        textProcessing:
          "$ref": "#/components/schemas/TextProcessingRecord"
        recognitionHints:
          "$ref": "#/components/schemas/ModelHintUsage"
        proofreadingHints:
          "$ref": "#/components/schemas/ModelHintUsage"
        formattingRejectionReason:
          type: "string"
        consumedListControls:
          type: "array"
          items:
            "$ref": "#/components/schemas/ListControlSpan"
        continuation:
          "$ref": "#/components/schemas/DictationContinuation"
        delivery:
          "$ref": "#/components/schemas/DeliveryReceipt"
        error:
          type: "string"
        progress:
          type: "number"
          minimum: 0
          maximum: 1
        importedSource:
          "$ref": "#/components/schemas/ImportedSource"
      required:
        - "schemaVersion"
        - "id"
        - "requestID"
        - "device"
        - "mode"
        - "status"
        - "createdAt"
        - "updatedAt"
        - "settings"
        - "rawText"
        - "finalText"
        - "insertionText"
        - "previewText"
    GenerationPage:
      type: "object"
      additionalProperties: false
      properties:
        items:
          type: "array"
          items:
            "$ref": "#/components/schemas/GenerationRecord"
        nextCursor:
          type: "string"
      required:
        - "items"
    WisprFlowImportResult:
      type: "object"
      additionalProperties: false
      properties:
        outcome:
          "$ref": "#/components/schemas/WisprFlowImportOutcome"
        record:
          "$ref": "#/components/schemas/GenerationRecord"
        unarchivedArtifactNames:
          type: "array"
          items:
            "$ref": "#/components/schemas/WisprFlowArtifactName"
      required:
        - "outcome"
        - "record"
        - "unarchivedArtifactNames"
    APIErrorResponse:
      type: "object"
      additionalProperties: false
      properties:
        code:
          type: "string"
        message:
          type: "string"
      required:
        - "code"
        - "message"
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      description: "Optional on loopback. Required for remote binding. Native clients send no Origin header."
  responses:
    APIError:
      description: "A request, admission, or server error."
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/APIErrorResponse"